Add functionality for env file to be saved locally, outside of project directory
